+++ app/Providers/Filament/AdminPanelProvider.php
@@ -30,6 +30,7 @@ class AdminPanelProvider extends PanelProvider
             ->path('/admin')
             ->viteTheme('resources/css/filament/admin/theme.css')
             ->login()
+            ->spa()
             ->tenant(Location::class)
             ->colors([
                 'primary' => Color::Amber,